# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: reviews
  - name: products
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List approved reviews
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a new review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [auto, moto, llanta]
        - name: brand
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    Review:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        vehicle:
          type: string
        text:
          type: string
        rating:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, location, vehicle, text, rating, createdAt]
    CreateReviewInput:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        vehicle:
          type: string
        text:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
      required: [name, location, vehicle, text, rating]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        category:
          type: string
          enum: [auto, moto, llanta]
        size:
          type: string
        price:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        inStock:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, brand, category, size, inStock, createdAt]
    ProductInput:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        category:
          type: string
          enum: [auto, moto, llanta]
        size:
          type: string
        price:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        inStock:
          type: boolean
      required: [name, brand, category, size, inStock]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]