# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from "./productCategory";

export interface Product {
  id: number;
  name: string;
  brand: string;
  category: ProductCategory;
  size: string;
  price?: number | null;
  description?: string | null;
  imageUrl?: string | null;
  inStock: boolean;
  createdAt: Date;
}