# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputCategory } from "./productInputCategory";

export interface ProductInput {
  name: string;
  brand: string;
  category: ProductInputCategory;
  size: string;
  price?: number | null;
  description?: string | null;
  imageUrl?: string | null;
  inStock: boolean;
}