# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  name: string;
  location: string;
  vehicle: string;
  text: string;
  rating: number;
  createdAt: Date;
}